pathmap: add definition for the display platform
Due to the nature of the aosp display source structure,
Android.mk files each have to check the display variant,
then return the correct platform name if the display is
aosp based.

add a definition so an Android.mk can easily find out
the platform of the device, be it msm8974, msm8x26,
msm8960, or '.' for caf based sources.

Change-Id: I87858c326bc256bed06ed58928d7a86ccf1bf9fa